Minor update
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@2421 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_array.c
@@ -1064,7 +1064,8 @@ mx_copy_array_to_buffer( void *array_pointer,
 		MX_DEBUG(-2,("%s: num_dimensions = 0, mx_datatype = %ld, "
 		"native_bytes_to_copy = %d, network_bytes_to_copy = %d",
 			fname, mx_datatype,
-			native_bytes_to_copy, network_bytes_to_copy));
+			(int) native_bytes_to_copy,
+			(int) network_bytes_to_copy));
 #endif
 
 		switch( mx_datatype ) {
